Pkg cleanup: moving out main function of danm to cmd/. 99% remained, renamed to metacni, and thus become testable
