+++ pkg/metacni/metacni_test.go
@@ -1,4 +1,4 @@
-package main
+package metacni
 
 import (
   "testing"
