+++ run_uts.sh
@@ -2,7 +2,7 @@
 set -e
 echo "" > coverage.out
 for d in $(go list ./... | grep -v vendor | grep -v crd); do
-    go test -covermode=count -v -coverprofile=profile.out -coverpkg=github.com/nokia/danm/pkg/cnidel,github.com/nokia/danm/pkg/bitarray,github.com/nokia/danm/pkg/ipam,github.com/nokia/danm/pkg/danmep,github.com/nokia/danm/pkg/danmnet,github.com/nokia/danm/pkg/syncher $d
+    go test -covermode=count -v -coverprofile=profile.out -coverpkg=github.com/nokia/danm/pkg/cnidel,github.com/nokia/danm/pkg/bitarray,github.com/nokia/danm/pkg/ipam,github.com/nokia/danm/pkg/danmep,github.com/nokia/danm/pkg/danmnet,github.com/nokia/danm/pkg/syncher,github.com/nokia/danm/pkg/metacni $d
     if [ -f profile.out ]; then
         cat profile.out >> coverage.out
         rm profile.out
